FE API KatuzSearch changed permission + KatuzId int64 to int32

+++ NOBY/NOBY.Api/Endpoints/Address/KatuzSearch/Dto/KatuzLine.cs
@@ -5,7 +5,7 @@ public class KatuzLine
     /// <summary>
     /// Id katastrálního území (KATUZ)
     /// </summary>
-    public long KatuzId { get; set; }
+    public int KatuzId { get; set; }
 
     /// <summary>
     /// Textová reprezentace katastrálního území
